refactor create-state into more modular structure
