+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { createState } from "./create-state";
 export { onMount, onUnmount } from "./lifecycle";
